Create License
This added the Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International license to the project. This means that anybody can remix this project but cannot use it for commercial use and must release it under the same terms if they do remix it. The details of the license can be found here: https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode.en